renamed TimePeriod => TimePeriodFilter and moved into types/filters.ts from TimePeriod.ts
