+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/types/TimePeriod.ts
@@ -1,21 +1,6 @@
-import { TimePeriodOption } from '@frontend/services/tableBuilderService';
 import { TimePeriodCode } from '@common/services/types/TimePeriod';
-import { Filter } from './filters';
 
-export default class TimePeriod extends Filter {
-  public readonly year: number;
-
-  public readonly code: string;
-
-  public constructor({ year, code, label }: TimePeriodOption) {
-    super({ label, value: `${year}_${code}` });
-
-    this.code = code;
-    this.year = year;
-  }
-}
-
-export function parseYearCodeTuple(value: string): [number, string] {
+export default function parseYearCodeTuple(value: string): [number, string] {
   const [year, code] = value.split('_');
 
   const parsedYear = Number(year);
