+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/types/filters.ts
@@ -1,6 +1,7 @@
 import {
   FilterOption,
   IndicatorOption,
+  TimePeriodOption,
 } from '@frontend/services/tableBuilderService';
 import camelCase from 'lodash/camelCase';
 
@@ -39,5 +40,18 @@ export class Indicator extends Filter {
   public constructor({ value, label, unit }: IndicatorOption) {
     super({ value, label });
     this.unit = unit;
+  }
+}
+
+export class TimePeriodFilter extends Filter {
+  public readonly year: number;
+
+  public readonly code: string;
+
+  public constructor({ year, code, label }: TimePeriodOption) {
+    super({ label, value: `${year}_${code}` });
+
+    this.code = code;
+    this.year = year;
   }
 }
